requirements.txt
discord.py\ncolorama
